euchre-api/requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiohttp-devtools==1.0.post0
aiosignal==1.2.0
anyio==3.6.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
devtools==0.9.0
executing==0.10.0
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
packaging==21.3
protobuf==4.21.9
Pygments==2.13.0
pyparsing==3.0.9
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.3.4
six==1.16.0
sniffio==1.3.0
watchgod==0.8.2
wrapt==1.14.1
yarl==1.8.1
